refactor: abstract isRouteActive into shared utility
