+++ frontend/web-app/src/components/navigation/sidebar-nav.tsx
@@ -13,24 +13,7 @@ import {
 } from "@tabler/icons-react";
 import { UserAvatar } from "../user/user-avatar";
 import { handleLogout } from "@/lib/Oauth-functions";
-// Helper function to check if a path matches a route pattern
-function isRouteActive(currentPath: string, routePattern: string): boolean {
-  // If it's an exact match
-  if (currentPath === routePattern) return true;
-
-  // If the route pattern ends with a dynamic segment (e.g., /messages/[id])
-  if (routePattern.includes("[") && routePattern.includes("]")) {
-    const basePath = routePattern.split("/[")[0];
-    return currentPath.startsWith(basePath);
-  }
-
-  // If the route pattern is a parent route (e.g., /messages)
-  if (currentPath.startsWith(routePattern + "/")) {
-    return true;
-  }
-
-  return false;
-}
+import { isRouteActive } from "@/lib/utils/route";
 
 export type SidebarNavItem = {
   title: string;
@@ -55,7 +38,7 @@ export function SidebarNav({
   return (
     <div
       className={cn(
-        "flex h-full flex-col justify-between  px-3 py-4 bg-background",
+        "hidden md:flex  h-full flex-col justify-between  px-3 py-4 bg-background",
         className
       )}
     >
